Remove hamsterjs from LayoutContainer.
It cannot detect when dom elements are overlapping, but not parented,
and the Correspondence maps are going to overlap large areas of the
BioMaps. So punt on the scroll wheel handler for now, although I think
hamsterjs will still be useful, in combination with a selection
tool. So leaving hamsterjs in the package.json.